requirements.txt
spacy==3.8.11
python-dotenv==1.2.2
groq==1.1.1
en-core-web-sm==3.8.0
wikipedia==1.4.0
